9/5 20:30 - 21.30 = 1h; continua gestione unità di misura, con lucio help sistemato database prima di dare fuoco e fiamme alla città.

+++ src/main/java/org/galvanica/model/DettaglioAlimentazione.java
@@ -16,7 +16,7 @@ public class DettaglioAlimentazione {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long idDettaglio;
     private String note;
-    private Double quantitaProdotto;
+    private Integer quantitaProdotto;
     @Enumerated(EnumType.STRING)
     private UnitaDiMisura unitaDiMisura;
     @ManyToOne
